[code-gen]: Models more prominently in FluentApp output
Made model a first class in app output, other plugins can reference it freely when needed.
This also means validator function name generation is happening in the generatorPlugin. For consistency, the routeTrie builder is also moved to the generatorPlugin.
For the Typescript builder it is more easy to generate all types now