code-gen: API client generator plugin
Generates an Axios client. Nothing fancy around it yet, which may come in the future

With that, a generateJsDoc function is abstracted and now available in the template context while generating. Currently resolves references directly, but that should change in the future once I've sorted out how that is going to work in combination with the default values and optional values