code-gen: refactor code-gen and group generator plugins
Refactored App to make it easier to use than the `runCodeGen().build()` variant. This app also supports extending the app with output of other apps. This ensures packages never have to depend on code-gen as a prod dependency.

The plugins are only moved to the new generator plugin system. In the future they should also support extending.

The plugin system is a combined system of the previous 'builder' and 'generator' plugins. The init's are merged, options are passed cleanly, etc.