code-gen: 5/x splitting types in plugins
More group base code generation. Which results in the validator and mocks exporting based on groups, which results in getting a better overview of what is generated.

Router now also uses a custom type plugin to define and store the routes. Which unified the generator plugins a bit more.

The model and mock plugin now check if a type is supported top level before generating a type / mock for the type