store: add a files abstraction
Currently supports creating, updating, copying and deleting for files. Updating is implicitly done when calling create with an existent id.
There is also support for getting (partial) streams for the files.

Not too much error handling is build in. Relying a bit on Minio for their validation.

Also added tests for migrations and minio

No support for multiple buckets yet.

CI is probably broken. Need to enable a postgres and minio service