code-gen: move more type and group related logic to App
Since we expect all users to use the predefined TypeBuilder and TypeCreator,
we can do a bunch of logic in the App.

So App now keeps a list of newly `add`-ed builders.
App also keeps the internal structure, instead of free for all.
App.extends will merge groups and items.
App stores in a grouped way.
App manages deduping of models.

This resulted in some logic from the former 'models' generator to move to App.

Also renamed 'models' generator to 'types' generator since that is what it is ;)

There is a bit more boilerplate now in the templates to only render a single group at a time, but it is probably more clear to read.