code-gen: add OpenAPI loader
Tried it on the Pet store swagger example, and on a internal document.
Seems to work ok. The biggest issue currently is the use of long descriptive names in some OpenAPI specs that I looked at. This meant that prettier will put for example the variable name on a new line in JSDoc which results in an invalid line and thus a 'useless' stub.