docs: .d.ts files and generated docs
This change introduces index.d.ts files per package. Currently only @lbu/cli and @lbu/insight are done yet. This results in a shift for the JSDoc usage. All types declared in a declaration file don't need to exist in the JSDoc. For exported functions this results in only basic usage of the `param` and `returns` tag, without any comments.

This also brings a docsify based docs folder. This allows for a neatly organized sidebar and search without doing special tricks.

From the index.d.ts file a api.md file is generated with `lbu typedoc`. This generates documentation for all modules that have a index.d.ts file and then concatenates that with `concat-md` into the api.md file.