cli: replace nodemon with chokidar
By removing nodemon we are removing a bunch of dependencies. We also don't need to rely on the `nodemonArgs` any more and provide a cleaner interface of ignoring files.

Uses tree-kill package for killing the spawned processes, since some OS's don't kill the whole process tree by default. We may vendor this in, but since it doesn't depend on anything else, this should be fine to use.

BREAKING CHANGE