code-gen: AppError and response validation in apiClient
If isNodeServer passed to App#generate is true the following things are generated in the apiClient:

- call to the appropriate response validator, to ensure the server does not leak any properties

- Errors are caught and converted to AppError

When validation errors are caught the key is prefixed with the group and name of the response type.
When server errors are caught, the key and info are picked from the response data and put in to a new AppError.
Other errors are converted to generic AppErrors putting the response and headers in the 'info' object.