code-gen: generate cancel token support in browser environments (#347)
We optional accept a `CancelToken.source()` in the generated `useQuery` hooks. We then automatically add `promise.cancel` and provide the token to the api client function. The apiClient now accepts optional `cancelToken`, which is only enabled currently in browser environments.

This still keeps the request cancellation manual, but provides greater flexibility, like cancelling multiple in flight requests with a single source, and doesn't create too many unnecessary / unreadable wraps in the apiClient

Closes #344