insight: improve performance of ndjson logging (#417)
We get some improvements by formatting the context once on logger creation.
Another small bump came from not creating extra objects, but doing the top level `JSON.stringify` inline.
The last improvement was to only call `stream.write` once, and probably the most notable improvement.

BREAKING CHANGE: The context is now nested in stead of spread on the top level log line. We also default to an empty object if the context is not specified.