code-gen: subquery support in 'in' and 'notIn' sql where
In preparation of whatever query builder we are going to create, this is most likely going to be a pattern that is used. For now, only the generated traverser uses this.

The type system does not yet allow external types, but if support for that is added, we should generate the correct XxxWhere type.