lint-config: use meriyah parser in prettier, disable eslint import rules
The meriyah parser is supported by Prettier since 2.2.0.
It strips 0.6 - 0.9 seconds of Prettier execution time on the lbu codebase.

This also disables some import and export related rules, that should just be found by editors, and saves a good 1.5s of ESLint execution time.

Total improvement on this codebase is around 20 %