code-gen: add dumpApiStructure and throwingValidators options
We only want to exposes the necessary things outside of the api. This is only the router and related types. This can now be enabled via `dumpApiStructure`.
This also results in `dumpStructure` never being enabled by default, and is thus manually needed when creating packages.

This also adds a `throwingValidators` option. This is expected by the router and sql generator. However it is only enabled by default with `isNodeServer`.

Closes #525