cli: fix watch and restart behaviour of long running processes
The 'close' listener did run for processes that were killed, introducing a race condition where the new instance was set to 'undefined' instead of the existing one.

We now wait on tree-kill before starting a new process, this fixes a race condition where a server may not be fully stopped yet, but a new process tries to listen on the same port.