chore: various code-gen OpenAPI fixex + bench using Github API spec (#609)
* code-gen: deep copy validation result on App extend

Several internal parts were broken when extend was used. The validators create objects without a prototype, which results in the `isPlainObject` checks reporting 'false' everywhere.

* code-gen: support default and optional path and query params in openAPI converter

* code-gen: transform named types in the OpenAPI importer

The github api uses dashes in their references

* bench: create benchmark against the Github v3 API

This is based on the OpenAPI importer. However it makes the whole dev env a bit slower since so many symbols are generated. This is directly a good test of scalability of the generators.