code-gen: fix name clash for self referencing tables in the query-builder
We solve this by generating an equivalent internal query-builder, that uses a different shortName when joining. This way we can hop back and forth between these functions if a nested self referencing query is executed.

The generator code didn't become any prettier, need to think a bit about it and get cleanup in at some point

Closes #611