cli: cleaner test output on failed results (#628)
Refactored the test printer a bit with more earlier returns and added some comments on why the order things is as is.

This also closes #624 by skipping sub trees even for test suites if nothing failed. We would print the first 2 levels of all trees when a failure happened, and the full sub tree to the specific failures. Now we only print the first single level of the tree and still do the full sub tree to a failure.

Fixes #624