ci: automate release process
When a commit on 'main' contains the words 'prepare' and 'release' this job should run. It will get the new version out of the latest commit message and execute Lerna.
After that we expect Lerna to have pushed some tags and fill in the Github release with the changelog for that version.