store: add stringifyQueryPart function (#648)
Helps with debugging queries or when you quickly want to see what query is generated. Supports a bit of interpolating variables, but may not work in a lot of cases. Haven't decided yet if I want to harden the implementation or remove it. But we'll roll with it for now

Closes #644