feat(store): handler timeout for job queue (#663)
Add the option to add a handlerTimeout on Queue worker creation. This defaults to 30 seconds. A faster timeout increases throughput and may force for smaller targeted jobs instead of doing all kinds of things in a single job. Combined with the retry count, this should prevent the worker from ever getting stuck on jobs that can't be completed.

Closes #651