fix(cli): proper exit on test worker failure
The test runner did not exit when a worker would fail in either the global setup or teardown execution.

Since we can't know when a message will arrive, the runner just waited till all workers posted their results back. A failing worker won't post anything, so it was an infinite timeout loop.

Know we exit as soon as possible when a worker exits with an exit code other than 0.

Closes #666