chore(ci): improve CI for external contributors
Push events are not triggered for forks, so we limit them to the `main` branch. For PR's we separately run the lint & test jobs, and make sure to not use the `GITHUB_TOKEN` secret if it is not available. This also means that we don't comment on the PR with bench & coverage results.