feat(code-gen): add sql order by support
Supports both a raw query part and array defining the sorted columns. If an array is passed, the `orderBySpec` argument is used to determine the order. Note that if no spec is given, the order is `ASC` as is the Postgres default.

When more custom ordering is necessary you should be able to just construct a query part, the `orderBySpec` and `tableName` are then ignored.

Closes #704