fix(code-gen): allow 'optional' order by for deletedAt column
Accidentally added the 'deletedAt' field to the exception list. That exception exists because the internal 'createdAt' & 'updatedAt' will never be 'NULL', so allowing 'ASC NULL LAST' doesn't have any effect on the output.

Also made the options named, which may result in a smaller validator output.

Removed the 'ASC NULLS LAST' & 'DESC NULLS FIRST' as they are the defaults for 'ASC' and 'DESC'