feat(code-gen): generate a code splittable api client
This is a breaking change. The `useApi` hook now stores an axios instance instead of the api client object. This also removes some of the 'isBrowser' checks that we had in here.
The generated react-query hooks also import the specific api functions, so pages should only include the api calls that they use in the bundles.

Refs #730