feat(server): support renewing sessions every x seconds (#1166)
Rolling sessions are just too quick as a default, and renew option is too slow. So we add a way in between.

If 'renew' is a number, it determines the amount of seconds between resetting the expire date with the 'maxAge'. It defaults to 30 seconds. Which results in each request after 30 seconds containing a 'Set-Cookie' header.