feat(store): add `unsafeIgnoreSorting` as a queue worker option (#1345)
References #1344

As the option implies this disables any sorting when picking up jobs. Which equals to ignoring the 'priority' and 'scheduledAt' properties of jobs. Resulting in a much faster way of picking up jobs, for example when generating bursts of a million jobs. Which often results in slow seq scans and in memory sorting by postgres.
A 'normal' workload consisting of transactional inserted jobs that are immediately picked up by Postgres should most likely use the existing indices and don't need this option.

Seeing improvements of 30 seconds to just a few milliseconds when picking up a job.