chore(tests): improve stability of SQL time-based tests
These tests failed if the Docker VM time was out of sync with the system time. Now they pass with a small margin of timing error. This doesn't improve the test reliability in the sense that the code now for sure works, but makes it more reliable for other contributors so their tests don't fail randomly.