feat(code-gen/experimental): redo type traversal
Previous variant where each type needed to reimplement the same logic was not a good idea. Was pretty happy with the latest type traverser that we had in current code-gen, so recreated that.

This means that a new type, without new structure behaviour can be simply added by defining it's traverse paths. All other stuff should just work.

Current implementation is not tested, so may have some bugs.

References #2010