feat(stdlib): completer Axios and Postgres error formatting by AppError#format
Added all possible fields of Postgres errors, a bunch of them will be undefined, which are skipped when serialized to JSON. When running in debug mode, the query & parameters will be added by the Postgres library, so these will be added as well.

The formatted Axios error is reshaped, adding request headers and preventing response streams from being added. The request headers are added and the Au

BREAKING CHANGE:
- `axios.requestPath` and `axios.requestMethod` are replaced by `axios.request.path` and `axios.request.method`.
- `axios.responseStatus`and `axios.responseBody` are replaced by `axios.response.status`, `axios.response.body`.
- `axios.responseHeaders` is removed from the error to prevent session tokens in the logs by default.