feat(code-gen/experimental): setup a behaviour / specification suite
This accomplice two goals:
- Make it easier to verify an implementation
- Support running the same tests for multiple generator targets.

This system is limited to testing errors and does not validate if the result is correct.

The experimental code-gen has to many breaking changes compared to legacy code-gen, so we don't include legacy in there.

This also means that `testExperimentalGenerateContext` and `testExperimentalGenerateFiles` should mostly be used to test preprocessor behaviour and specific outputs.