feat(code-gen/experimental): support database.includeEntityDiagram
This writes a Markdown file with a Mermaid code block visualising the ERD.

Compared to `compas visualise` that we use in legacy code-gen, this doesn't require an
external system dependency. That system dependency causes issues when a there is quite
a difference in versions available in the different system package managers which all have different outputs, causing unwieldy diffs.