feat(store): improve error handling around migrations
The Postgres error information is already included in when `AppError#format` recursively calls itself for `e.cause`, so we don't have to repeat that.

Also, use a normal transaction function to make it easier to reason about the order in which the migration source is run and when the migration is inserted.