feat(compas): add (nested) config loading
This config file is located at `config/compas.json`. Opted to not use the existing config loaders in Compas. We want to cache the resolved config, which involves a few extra steps if we need to cache an object that allows functions etc. May want to revisit this later, depending on the complexity of the config object.