feat(compas): start with cache
As anything else, this is pretty bare for now. We load the cache file from disk and able to write the file to disk again. Opted to validate the cache so only known things can be written to the cache, and manual tampering is ignored.

Intentionally not writing cache information to the tui. This should be an implementation detail and everything should work fast if the cache is failing.