fix(server): add explicit resource cleanup in the wrapped response streams
The original `.on("error")` prevented Koa from doing its magic to automatically destroy the source stream when the client closed the connection. This should properly fix it.

In Koa v4 (at some point) we may be able to use the Node.js provided stream.pipeline, depends a bit on how the changes shape up. Without having to override if we send a response or if Koa does it.